Первый этап интеграции с веб-интерфейсом

+++ mini_server/SocketFactory.cpp
@@ -1,21 +1,51 @@
+#include <sys/un.h>
+#include <stdexcept>
+#include <sys/stat.h>
+#include <format>
+#include <libgen.h>
+
 #include "SocketFactory.h"
 
 using namespace mini_server;
 
 extern "C++" int SocketFactory::createListeningSocket(const std::string &name, int maxConnections) {
+    struct stat socketStat{};
+
+    char fileName[name.size() + 1];
+    name.copy(fileName, sizeof (fileName));
+
+    auto dir = dirname(fileName);
+
+    if (stat(dir, &socketStat) != 0) {
+        if (errno == ENOENT) {
+            mkdir(dir, 0755);
+        }
+        if (errno != ENOENT && S_ISDIR(socketStat.st_mode)) {
+            throw std::runtime_error(std::format("{} should be a directory for {} {}", dir, name, strerror(errno)));
+        }
+    }
+
+    if (stat(name.c_str(), &socketStat) != 0 && errno != ENOENT) {
+        throw std::runtime_error(std::format("Stat error for file {}, {}", name, strerror(errno)));
+    }
+
+    if (errno != ENOENT && S_ISFIFO(socketStat.st_mode)) {
+        throw std::runtime_error(std::format("Unlink error for file {}, {}. File should be fifo or can be deleted", name, strerror(errno)));
+    }
+
     auto sock = socket(AF_UNIX, SOCK_STREAM, 0);
 
     if (sock < 0) {
         throw std::runtime_error(strerror(errno));
     }
 
-    sockaddr addr = {AF_UNIX, ""};
+    sockaddr_un addr = {AF_UNIX, ""};
 
-    name.copy(addr.sa_data, sizeof(addr.sa_data), 0);
+    name.copy(addr.sun_path, sizeof(addr.sun_path), 0);
 
     unlink(name.c_str());
 
-    if (bind(sock, &addr, sizeof(addr)) < 0) {
+    if (bind(sock, (sockaddr *)&addr, sizeof(addr)) < 0) {
         throw std::runtime_error(strerror(errno));
     }
